Live_call/audio through websocket/server/requirements.txt
torch>=2.0.0
openai-whisper
numpy
soundfile
websockets
